requirements.txt
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
# dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-import-export==3.3.3
django-jazzmin==2.6.0
et-xmlfile==1.1.0
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
Pillow==10.1.0
# psycopg2==2.9.9
PyMySQL==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
